refactor: split out prebuild codes
