--- app/routes/_root/types.ts
@@ -1,15 +1,0 @@
-interface MenuDocAttributes {
-  title: string
-  order?: number
-  new?: boolean
-  [key: string]: unknown
-}
-
-export interface MenuDoc {
-  attrs: MenuDocAttributes
-  parentSlug?: string
-  children: MenuDoc[]
-  filename: string
-  hasContent: boolean
-  slug: string
-}
