+++ scripts/services/md-plugins/strip-link-ext.ts
